+++ src/App.jsx
@@ -1,6 +1,6 @@
 import "./App.css";
 import { Outlet } from "react-router-dom";
-import Header from "./components/header/Header";
+import Header from "./components/Header/Header";
 import Footer from "./components/Footer/Footer";
 import { useLocation } from "react-router-dom";
 import { AuthProvider } from "./utils/authContext";
